requirements.txt
PyQt5~=5.15.1
opencv-python~=4.4.0.42
numpy~=1.19.1
pandas~=1.1.4
qimage2ndarray~=1.8.3
Pillow~=7.2.0
rich~=9.12.0
pytesseract~=0.3.6
pywin32~=228
openpyxl~=3.0.5
argparse~=1.4.0